Refactor Scanner, Add ScannerFactory
